requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
playwright==1.42.0
pandas==2.2.1
python-multipart==0.0.9
python-dotenv